Keep IME regression coverage under file-length guard
The workflow guard failed because the new IME regression and dedup helper grew two Swift files that are already tracked by the line-length budget. Move the regression into its own XCTest file and extract the dedup/text filtering support into a small production helper so the behavior remains covered without accepting more debt in those files.

Constraint: workflow-guard-tests rejects growth in tracked Swift files over budget
Rejected: Refresh file-length budget | would accept avoidable debt instead of reducing growth
Confidence: high
Scope-risk: narrow
Directive: Keep issue #3284 coverage executable; do not replace it with source-shape assertions
Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv
Tested: CMUX_SKIP_ZIG_BUILD=1 ./scripts/reload.sh --tag issue-3284-numpad-ime-duplicate
Not-tested: App-hosted cmux-unit run, blocked locally before this CI-fix pass by unrelated SysDefined-event assertion before selected XCTest
