+++ Sources/GhosttyTerminalView.swift
@@ -6910,17 +6910,11 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
     }
 
     // For NSTextInputClient - accumulates text during key events
-    private var keyTextAccumulator: [String]? = nil
+    private(set) var keyTextAccumulator: [String]? = nil
     private var markedText = NSMutableAttributedString()
     private var lastPerformKeyEvent: TimeInterval?
-    private var externalCommittedTextDepth = 0
-    private struct PendingNumpadIMECommitDedup {
-        let text: String
-        let keyCode: UInt16
-        let sourceId: String?
-        let timestamp: TimeInterval
-    }
-    private var pendingNumpadIMECommitDedup: PendingNumpadIMECommitDedup?
+    private(set) var externalCommittedTextDepth = 0
+    var numpadIMECommitDeduplicator = NumpadIMECommitDeduplicator()
     private struct SelectionSnapshot {
         let range: NSRange
         let string: String
@@ -7807,65 +7801,6 @@ class GhosttyNSView: NSView, NSUserInterfaceValidations {
         guard let chars = (event.characters(byApplyingModifiers: []) ?? event.charactersIgnoringModifiers ?? event.characters),
               let scalar = chars.unicodeScalars.first else { return 0 }
         return scalar.value
-    }
-
-    private func isControlCharacterScalar(_ scalar: UnicodeScalar) -> Bool {
-        scalar.value < 0x20 || scalar.value == 0x7F
-    }
-
-    private func shouldSendText(_ text: String) -> Bool {
-        guard !text.isEmpty else { return false }
-        if text.count == 1, let scalar = text.unicodeScalars.first {
-            return !isControlCharacterScalar(scalar)
-        }
-        return true
-    }
-
-    private func notePotentialDeferredNumpadIMECommit(text: String, event: NSEvent) {
-        let flags = event.modifierFlags
-            .intersection(.deviceIndependentFlagsMask)
-            .subtracting([.function, .capsLock])
-        guard flags == [.numericPad], text.allSatisfy(\.isNumber) else {
-            pendingNumpadIMECommitDedup = nil
-            return
-        }
-
-        pendingNumpadIMECommitDedup = PendingNumpadIMECommitDedup(
-            text: text,
-            keyCode: event.keyCode,
-            sourceId: KeyboardLayout.id,
-            timestamp: ProcessInfo.processInfo.systemUptime
-        )
-    }
-
-    private func shouldSuppressDeferredNumpadIMECommit(_ text: String) -> Bool {
-        guard externalCommittedTextDepth == 0,
-              keyTextAccumulator == nil,
-              let pending = pendingNumpadIMECommitDedup else {
-            return false
-        }
-
-        let now = ProcessInfo.processInfo.systemUptime
-        guard now - pending.timestamp <= 0.25 else {
-            pendingNumpadIMECommitDedup = nil
-            return false
-        }
-
-        guard pending.text == text, pending.sourceId == KeyboardLayout.id else {
-            return false
-        }
-
-        if let currentEvent = NSApp.currentEvent, currentEvent.type == .keyDown {
-            let currentFlags = currentEvent.modifierFlags
-                .intersection(.deviceIndependentFlagsMask)
-                .subtracting([.function, .capsLock])
-            guard currentFlags == [.numericPad], currentEvent.keyCode == pending.keyCode else {
-                return false
-            }
-        }
-
-        pendingNumpadIMECommitDedup = nil
-        return true
     }
 
     /// If AppKit consumed Shift+Space for IME/input-source switching, interpretKeyEvents
