+++ cmuxTests/CJKIMEInputTests.swift
@@ -9,7 +9,7 @@ import ObjectiveC.runtime
 #endif
 
 private var cjkIMEInterpretKeyEventsSwizzled = false
-private var cjkIMEInterpretKeyEventsHook: ((GhosttyNSView, [NSEvent]) -> Bool)?
+var cjkIMEInterpretKeyEventsHook: ((GhosttyNSView, [NSEvent]) -> Bool)?
 private var ghosttyPasteActionSwizzled = false
 private var ghosttyPasteActionHook: ((GhosttyNSView, Any?) -> Void)?
 private var ghosttyPasteAsPlainTextActionSwizzled = false
@@ -34,7 +34,7 @@ private extension GhosttyNSView {
     }
 }
 
-private func installCJKIMEInterpretKeyEventsSwizzle() {
+func installCJKIMEInterpretKeyEventsSwizzle() {
     guard !cjkIMEInterpretKeyEventsSwizzled else { return }
 
     let originalSelector = #selector(GhosttyNSView.interpretKeyEvents(_:))
@@ -128,7 +128,7 @@ private func installGhosttyPasteActionSwizzle() {
     ghosttyPasteAsPlainTextActionSwizzled = true
 }
 
-private func findGhosttyNSView(in view: NSView) -> GhosttyNSView? {
+func findGhosttyNSView(in view: NSView) -> GhosttyNSView? {
     if let view = view as? GhosttyNSView {
         return view
     }
@@ -1296,117 +1296,6 @@ final class KoreanIMEMarkedTextLeakRegressionTests: XCTestCase {
         XCTAssertTrue(capturedEvent.composing, "Hangul composition keyDown should stay in composing mode")
         XCTAssertNil(capturedEvent.text, "Uncommitted Hangul jamo must not be encoded into the terminal surface")
         XCTAssertTrue(view.hasMarkedText(), "Composition should remain active until the IME commits or cancels")
-    }
-}
-
-@MainActor
-final class TraditionalChineseIMENumpadRegressionTests: XCTestCase {
-    private struct HostedTerminalWindow {
-        let surface: TerminalSurface
-        let window: NSWindow
-        let hostedView: GhosttySurfaceScrollView
-        let surfaceView: GhosttyNSView
-    }
-
-    private func makeHostedTerminalWindow() throws -> HostedTerminalWindow {
-        _ = NSApplication.shared
-
-        let surface = TerminalSurface(
-            tabId: UUID(),
-            context: GHOSTTY_SURFACE_CONTEXT_SPLIT,
-            configTemplate: nil,
-            workingDirectory: nil
-        )
-        let hostedView = surface.hostedView
-
-        let window = NSWindow(
-            contentRect: NSRect(x: 0, y: 0, width: 360, height: 240),
-            styleMask: [.titled, .closable],
-            backing: .buffered,
-            defer: false
-        )
-
-        let contentView = try XCTUnwrap(window.contentView)
-        hostedView.frame = contentView.bounds
-        hostedView.autoresizingMask = [.width, .height]
-        contentView.addSubview(hostedView)
-
-        window.makeKeyAndOrderFront(nil)
-        window.displayIfNeeded()
-        contentView.layoutSubtreeIfNeeded()
-        hostedView.setVisibleInUI(true)
-        hostedView.setActive(true)
-        RunLoop.current.run(until: Date().addingTimeInterval(0.05))
-
-        let surfaceView = try XCTUnwrap(findGhosttyNSView(in: hostedView))
-        return HostedTerminalWindow(
-            surface: surface,
-            window: window,
-            hostedView: hostedView,
-            surfaceView: surfaceView
-        )
-    }
-
-    func testKeypadDigitDoesNotDuplicateWhenTraditionalChineseIMECommitsAfterKeyDown() throws {
-        let hostedTerminal = try makeHostedTerminalWindow()
-        let terminalSurface = hostedTerminal.surface
-        let window = hostedTerminal.window
-        let surfaceView = hostedTerminal.surfaceView
-        let previousKeyEventObserver = GhosttyNSView.debugGhosttySurfaceKeyEventObserver
-        let previousInputSourceOverride = KeyboardLayout.debugInputSourceIdOverride
-        let previousInterpretHook = cjkIMEInterpretKeyEventsHook
-        defer {
-            GhosttyNSView.debugGhosttySurfaceKeyEventObserver = previousKeyEventObserver
-            KeyboardLayout.debugInputSourceIdOverride = previousInputSourceOverride
-            cjkIMEInterpretKeyEventsHook = previousInterpretHook
-            window.orderOut(nil)
-            withExtendedLifetime(terminalSurface) {}
-        }
-
-        KeyboardLayout.debugInputSourceIdOverride = "com.apple.inputmethod.TCIM.Pinyin"
-        installCJKIMEInterpretKeyEventsSwizzle()
-        cjkIMEInterpretKeyEventsHook = { candidateView, _ in
-            guard candidateView === surfaceView else { return false }
-            DispatchQueue.main.async {
-                candidateView.insertText("1", replacementRange: NSRange(location: NSNotFound, length: 0))
-            }
-            return true
-        }
-
-        var pressedText: [String] = []
-        GhosttyNSView.debugGhosttySurfaceKeyEventObserver = { keyEvent in
-            previousKeyEventObserver?(keyEvent)
-            guard keyEvent.action == GHOSTTY_ACTION_PRESS, let text = keyEvent.text else { return }
-            pressedText.append(String(cString: text))
-        }
-
-        guard let event = NSEvent.keyEvent(
-            with: .keyDown,
-            location: .zero,
-            modifierFlags: [.numericPad],
-            timestamp: ProcessInfo.processInfo.systemUptime,
-            windowNumber: window.windowNumber,
-            context: nil,
-            characters: "1",
-            charactersIgnoringModifiers: "1",
-            isARepeat: false,
-            keyCode: 83
-        ) else {
-            XCTFail("Failed to create keypad 1 event")
-            return
-        }
-
-        window.makeFirstResponder(surfaceView)
-        withExtendedLifetime(terminalSurface) {
-            surfaceView.keyDown(with: event)
-            RunLoop.current.run(until: Date().addingTimeInterval(0.05))
-        }
-
-        XCTAssertEqual(
-            pressedText,
-            ["1"],
-            "Traditional Chinese IME numpad commits should not double-dispatch a keypad digit after keyDown fallback"
-        )
     }
 }
